+++ packages/serverless-components/aws-iam-role/CHANGELOG.md
@@ -3,7 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [1.2.0-alpha.15](https://github.com/getjerry/serverless-next.js/compare/@getjerry/aws-iam-role@1.1.12...@getjerry/aws-iam-role@1.2.0-alpha.15) (2023-10-06)
+# [1.2.0-alpha.17](https://github.com/getjerry/serverless-next.js/compare/@getjerry/aws-iam-role@1.2.0-alpha.13...@getjerry/aws-iam-role@1.2.0-alpha.17) (2023-10-06)
+
+**Note:** Version bump only for package @getjerry/aws-iam-role
+
+# [1.2.0-alpha.13](https://github.com/getjerry/serverless-next.js/compare/@getjerry/aws-iam-role@1.1.12...@getjerry/aws-iam-role@1.2.0-alpha.13) (2023-10-05)
 
 ### Bug Fixes