+++ packages/serverless-components/domain/CHANGELOG.md
@@ -3,7 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [1.7.0-alpha.29](https://github.com/getjerry/serverless-next.js/compare/@getjerry/domain@1.1.11...@getjerry/domain@1.7.0-alpha.29) (2023-10-06)
+# [1.7.0-alpha.31](https://github.com/getjerry/serverless-next.js/compare/@getjerry/domain@1.7.0-alpha.27...@getjerry/domain@1.7.0-alpha.31) (2023-10-06)
+
+**Note:** Version bump only for package @getjerry/domain
+
+# [1.7.0-alpha.27](https://github.com/getjerry/serverless-next.js/compare/@getjerry/domain@1.1.11...@getjerry/domain@1.7.0-alpha.27) (2023-10-05)
 
 ### Bug Fixes