+++ packages/serverless-components/nextjs-component/CHANGELOG.md
@@ -3,11 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [2.9.0-alpha.127](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@2.9.0-alpha.126...@getjerry/serverless-next@2.9.0-alpha.127) (2023-10-06)
+# [2.9.0-alpha.130](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@2.9.0-alpha.129...@getjerry/serverless-next@2.9.0-alpha.130) (2023-10-09)
 
 **Note:** Version bump only for package @getjerry/serverless-next
 
-# [2.9.0-alpha.126](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@2.0.5-alpha.6...@getjerry/serverless-next@2.9.0-alpha.126) (2023-10-06)
+# [2.9.0-alpha.129](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@2.9.0-alpha.124...@getjerry/serverless-next@2.9.0-alpha.129) (2023-10-06)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+# [2.9.0-alpha.124](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@2.9.0-alpha.123...@getjerry/serverless-next@2.9.0-alpha.124) (2023-10-05)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+# [2.9.0-alpha.123](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@2.9.0-alpha.122...@getjerry/serverless-next@2.9.0-alpha.123) (2023-10-05)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+# [2.9.0-alpha.122](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@2.0.5-alpha.6...@getjerry/serverless-next@2.9.0-alpha.122) (2023-10-05)
 
 ### Bug Fixes